Add impedance and wake function for multicell RF cavities
Impedance and wake formulas for an array of pillbox cavities.
The formula are from K. Bane "Short-range dipole wakefields
in accelerating structures for the NLC", SLAC-PUB-9963 (2003).

The formula is valid for the short range wake (e.g high frequency
impedance), as stated in the paper (criterion s/L < 0.15, with s the
wake range and L the length of the array of pillboxes).
It is up to the user to check the time/frequency range for which
he applies the function.