fix: complete component check in simulation folder when wake is computed
When the wake is computed with IW2D, we must add the wake components
wxdip, wxqua etc. and the wycst and zycst only in the case of a
FlatChamber computation.

Before the fix all possible components were added (indluding wxcst and
zxcst). This caused the simulations to be relaunched every time we tried
to retrieve the impedance/wake model, since some components appeared to
be missing from the simulation results folder.