fix: retrive IW2D _precise output files when the wake is computed
IW2D provides a second set of output files when the wake is computed.
These files have double the number of mesh points (for the wake only)
and have a _precise suffix.

When retrieving the impedance and wake with pywit, this caused an
AssertionError because two files in the same folder correspond to the
same plane and impedance-wake component.

Now we provide import_data_iw2d function common_string='_precise' if the
simulations were including the wake, such that we retrieve the data with
more points.