fix: complete test functions in test_interface.py
We add several test case to the test_check_already_computed function
from test_interface.py to check that: -computation is relaunched if a
component is missing in the output folder - computation is launched if
the folder structure is missing.

The iw2d input fixture is also reworked to be parameteric: it takes a
dictionnary as an input, with keys for the wake computation and the
chamber type.